+++ src/main/java/com/t1/snezhko/task1/aop/annotations/Cached.java
@@ -1,0 +1,11 @@
+package com.t1.snezhko.task1.aop.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface Cached {
+}
