+++ src/main/java/com/t1/snezhko/task1/cache/CacheConfig.java
@@ -1,0 +1,24 @@
+package com.t1.snezhko.task1.cache;
+
+import com.github.benmanes.caffeine.cache.Cache;
+import com.github.benmanes.caffeine.cache.Caffeine;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.util.concurrent.TimeUnit;
+
+@Configuration
+public class CacheConfig {
+
+    @Value("${app.cache.ttl}")
+    private int ttl;
+
+    @Bean
+    public Cache<String, Object> ttlCache() {
+        return Caffeine.newBuilder()
+                .expireAfterWrite(ttl, TimeUnit.SECONDS)
+                .maximumSize(5)
+                .build();
+    }
+}
