+++ src/main/java/com/t1/snezhko/task1/errorlog/persistence/repository/DataSourceErrorLogRepository.java
@@ -1,5 +1,6 @@
 package com.t1.snezhko.task1.errorlog.persistence.repository;
 
+import com.t1.snezhko.task1.aop.annotations.Cached;
 import com.t1.snezhko.task1.errorlog.persistence.entity.DataSourceErrorLogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
