+++ src/main/java/com/t1/snezhko/task1/measure/repositories/TimeLimitExceedLogRepository.java
@@ -1,5 +1,6 @@
 package com.t1.snezhko.task1.measure.repositories;
 
+import com.t1.snezhko.task1.aop.annotations.Cached;
 import com.t1.snezhko.task1.measure.entities.TimeLimitExceedLogEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
