+++ src/main/java/com/t1/snezhko/task1/transaction/persistence/repositories/TransactionRepository.java
@@ -1,5 +1,6 @@
 package com.t1.snezhko.task1.transaction.persistence.repositories;
 
+import com.t1.snezhko.task1.aop.annotations.Cached;
 import com.t1.snezhko.task1.transaction.persistence.entity.TransactionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
