+++ src/main/resources/application.properties
@@ -8,3 +8,4 @@ spring.liquibase.enabled=true
 spring.liquibase.change-log=classpath:db/changelog/changelog-master.xml
 
 app.method_execution.time_limit=150
+app.cache.ttl=10
